+++ src/components/SearchBar/SearchBar.vue
@@ -9,20 +9,22 @@
       class="min-w-full"
       :label="t('search.textLabel')"
       :sr-only-label="true"
-      input-class="focus-within:bg-white-300"
     >
       <template #iconLeft>
-        <Search class="w-6 h-6" />
+        <MagnifyingGlass size="l" />
       </template>
       <template #iconRight>
         <button
-          class="block"
+          :class="[
+            'block',
+            searchTerm ? 'opacity-100 cursor-pointer' : 'opacity-0'
+          ]"
           :aria-label="t('search.closeLabel')"
           :disabled="disabled"
           data-testid="clearSearchButton"
           @click="clearSearch"
         >
-          <Close class="w-6 h-6" />
+          <XmarkLarge class="w-6 h-6" />
         </button>
       </template>
     </text-input>
@@ -78,12 +80,12 @@ import LobTable from '../Table/Table';
 import TableBody from '../Table/TableBody';
 import TableRow from '../Table/TableRow';
 import LobLink from '../Link/Link.vue';
-import Search from '../Icons/Search';
-import Close from '../Icons/Close';
+import MagnifyingGlass from '../Icons/MagnifyingGlass';
+import XmarkLarge from '../Icons/XmarkLarge';
 
 export default {
   name: 'SearchBar',
-  components: { TextInput, LobTable, TableBody, TableRow, LobLink, Search, Close },
+  components: { TextInput, LobTable, TableBody, TableRow, LobLink, MagnifyingGlass, XmarkLarge },
   props: {
     searchFunction: {
       type: Function,
@@ -150,8 +152,10 @@ export default {
       });
     },
     clearSearch () {
-      this.searchTerm = '';
-      this.searchResults = [];
+      if (this.searchTerm) {
+        this.searchTerm = '';
+        this.searchResults = [];
+      }
     },
     onClickOutside ($event) {
       if (this.$refs.searchBar) {