+++ src/theme/icons/Icons.vue
@@ -1,14 +1,17 @@
 <template>
-  <div class="grid grid-cols-7 gap-2 text-black">
+  <div class="w-full grid gap-3 text-black">
     <div
       v-for="icon in icons"
       :key="icon"
       class="p-2"
     >
-      <span class="text-xs">{{ icon.name }}</span>
+      <div class="text-xs mb-1.5">
+        {{ icon.name }}
+      </div>
       <component
         :is="icon"
-        class="w-6 h-6"
+        :size="size"
+        :class="`text-${storyOnlyColorClass}`"
       />
     </div>
   </div>
@@ -20,6 +23,19 @@ import * as Icons from '@/components/Icons';
 export default {
   name: 'Icons',
   components: { ...Icons },
+  props: {
+    size: {
+      type: String,
+      default: 'm',
+      validator: function (value) {
+        return ['xxl', 'xl', 'l', 'm', 's'].includes(value);
+      }
+    },
+    storyOnlyColorClass: {
+      type: String,
+      default: 'black'
+    }
+  },
   computed: {
     icons () {
       return Icons;
@@ -27,3 +43,9 @@ export default {
   }
 };
 </script>
+
+<style scoped>
+.grid {
+  grid-template-columns: repeat(auto-fit, minmax(120px, 1fr));
+}
+</style>